Rewrite playbook to target the new Docker Registry 2.0
Initially, all the old ansible files are removed as the role is converted
over to using the new registry and the corresponding new configuration
settings. This was done to eliminate the complexity of getting things up
and running initially, and the elements that have been removed can be
re-introduced one by one to ensure they work with the new registry.

For this commit, the registry has no custom configuration settings and runs
simply as a local registry with images stored inside the container. Gradually,
this will be built back up to using an S3 storage system and an nginx proxy
for authentication.