Add nginx proxy configuration to handle registry authentication
This commit adds a 'proxy' play with tasks to configure and run an nginx
instance within a docker container, linked to the docker registry container,
in order to allow unauthenticated reads, but require authentication for
writes. Since this is not possible using the built-in registry authentication
system, the nginx solution was necessary.

In addition to setting up the proxy tasks, many previously un-configurable
values were converted into variables for flexibility and to ensure the values
are defined in a single place.

For testing purposes, a test.crt and .key are provided in the /tests/secrets
directory, and the test playbook was updated accordingly.

Currently, there remains an issue with nginx authentication and the 'docker
login' command. Because the docker login expects a 403 challenge when attempting
to log in (with a GET request), and nginx is configured not to issue such a
challenge on GETs, the docker client fails to authenticate properly. However,
nginx is still preventing unauthenticated writes, so another solution needs to
be found to resolve this.