Update nginx configuration to require authentication on all requests
Originally, we had planned to allow unauthenticated reads but have pushes
require a valid user/pass. Unfortunately, due to several quirks in how the
new docker registry and client are implemented, this is not possible:

  - https://github.com/docker/distribution/issues/1028
  - https://github.com/docker/docker/issues/17317

Because of this, we are falling back to requiring all operations to be
authenticated. Rather than hard-coding a list of users, the htpasswd task
that renders the login information is generated via an ansible variable
containing an array of user/pass combinations. This will allow flexibility
going forward should we want to be more granular in how we give access to
the registry. Initially, there will just be a 'root' account for systems like
Jenkins to use, and a 'dev' account for developers to use.

While this decision does obviate the need for an nginx proxy (since we could
just use the built-in registry authentication), we are leaving the proxy in
place in the hopes that the above issues are resolved soon and we can go back
to configuring nginx to allow anonymous reads.