Add debug server and health checks to registry configuration
This commit enables the built-in debug server as well as health checks
for the configured storage backend. This enables a periodic ping that
will check to see if the storage backend is healthy, and will deposit the
results of the health checks at the /debug/health endpoint of the debug
server.

The server itself runs on port 5001 in the registry container, but is
exposed via port 8080 on the nginx side. The same authentication is in place
for the debug server as for the registry itself.

More info can be found here:

  - https://github.com/docker/distribution/blob/master/docs/configuration.md#health